+++ pkg/logql/syntax/ast_test.go
@@ -463,26 +463,30 @@ func BenchmarkContainsFilter(b *testing.B) {
 
 func Test_parserExpr_Parser(t *testing.T) {
 	tests := []struct {
-		name    string
-		op      string
-		param   string
-		want    log.Stage
-		wantErr bool
+		name      string
+		op        string
+		param     string
+		want      log.Stage
+		wantErr   bool
+		wantPanic bool
 	}{
-		{"json", OpParserTypeJSON, "", log.NewJSONParser(), false},
-		{"unpack", OpParserTypeUnpack, "", log.NewUnpackParser(), false},
-		{"logfmt", OpParserTypeLogfmt, "", log.NewLogfmtParser(), false},
-		{"pattern", OpParserTypePattern, "<foo> bar <buzz>", mustNewPatternParser("<foo> bar <buzz>"), false},
-		{"pattern err", OpParserTypePattern, "bar", nil, true},
-		{"regexp", OpParserTypeRegexp, "(?P<foo>foo)", mustNewRegexParser("(?P<foo>foo)"), false},
-		{"regexp err ", OpParserTypeRegexp, "foo", nil, true},
+		{"json", OpParserTypeJSON, "", log.NewJSONParser(), false, false},
+		{"unpack", OpParserTypeUnpack, "", log.NewUnpackParser(), false, false},
+		{"logfmt", OpParserTypeLogfmt, "", log.NewLogfmtParser(), false, false},
+		{"pattern", OpParserTypePattern, "<foo> bar <buzz>", mustNewPatternParser("<foo> bar <buzz>"), false, false},
+		{"pattern err", OpParserTypePattern, "bar", nil, true, true},
+		{"regexp", OpParserTypeRegexp, "(?P<foo>foo)", mustNewRegexParser("(?P<foo>foo)"), false, false},
+		{"regexp err ", OpParserTypeRegexp, "foo", nil, true, true},
+		{"unknown op", "DummyOp", "", nil, true, false},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			e := &LabelParserExpr{
-				Op:    tt.op,
-				Param: tt.param,
+			var e *LabelParserExpr
+			if tt.wantPanic {
+				require.Panics(t, func() { e = newLabelParserExpr(tt.op, tt.param) })
+				return
 			}
+			e = newLabelParserExpr(tt.op, tt.param)
 			got, err := e.Stage()
 			if (err != nil) != tt.wantErr {
 				t.Errorf("parserExpr.Parser() error = %v, wantErr %v", err, tt.wantErr)
@@ -493,6 +497,32 @@ func Test_parserExpr_Parser(t *testing.T) {
 			} else {
 				require.Equal(t, tt.want, got)
 			}
+		})
+	}
+}
+
+func Test_parserExpr_String(t *testing.T) {
+	tests := []struct {
+		name  string
+		op    string
+		param string
+		want  string
+	}{
+		{"valid regexp", OpParserTypeRegexp, "foo", `| regexp "foo"`},
+		{"empty regexp", OpParserTypeRegexp, "", `| regexp ""`},
+		{"valid pattern", OpParserTypePattern, "buzz", `| pattern "buzz"`},
+		{"empty pattern", OpParserTypePattern, "", `| pattern ""`},
+		{"valid logfmt", OpParserTypeLogfmt, "", `| logfmt`},
+		{"valid json", OpParserTypeJSON, "", `| json`},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			l := LabelParserExpr{
+				Op:    tt.op,
+				Param: tt.param,
+			}
+			require.Equal(t, tt.want, l.String())
 		})
 	}
 }