+++ pkg/storage/stores/indexshipper/compactor/deletion/delete_requests_manager_test.go
@@ -679,6 +679,8 @@ type mockDeleteRequestsStore struct {
 	getAllUser   string
 	getAllResult []DeleteRequest
 	getAllErr    error
+
+	genNumber string
 }
 
 func (m *mockDeleteRequestsStore) GetDeleteRequestsByStatus(_ context.Context, _ DeleteRequestStatus) ([]DeleteRequest, error) {
@@ -707,4 +709,8 @@ func (m *mockDeleteRequestsStore) GetDeleteRequestGroup(ctx context.Context, use
 func (m *mockDeleteRequestsStore) GetAllDeleteRequestsForUser(ctx context.Context, userID string) ([]DeleteRequest, error) {
 	m.getAllUser = userID
 	return m.getAllResult, m.getAllErr
+}
+
+func (m *mockDeleteRequestsStore) GetCacheGenerationNumber(ctx context.Context, userID string) (string, error) {
+	return m.genNumber, m.getErr
 }