+++ pkg/util/marshal/marshal.go
@@ -3,6 +3,7 @@
 package marshal
 
 import (
+	"fmt"
 	"io"
 
 	"github.com/gorilla/websocket"
@@ -20,7 +21,7 @@ import (
 func WriteQueryResponseJSON(v logqlmodel.Result, w io.Writer) error {
 	value, err := NewResultValue(v.Data)
 	if err != nil {
-		return err
+		return fmt.Errorf("could not write JSON response: %w", err)
 	}
 
 	q := loghttp.QueryResponse{